+++ bms/src/com/raceup/ed/bms/gui/frame/data/Cell.java
@@ -20,28 +20,20 @@ import com.raceup.ed.bms.gui.frame.chart.ChartFrame;
 
 import javax.swing.*;
 import java.awt.*;
-import java.text.DecimalFormat;
 
 /**
  * GUI frame that contains battery cell info
  */
-class Cell extends JPanel {
-    private static final DecimalFormat SHORT_DEC_FORMAT = new DecimalFormat
-            ("0000.00");  // decimal places
-    private static final Color VALUE_TOO_HIGH_COLOR = Color.RED;
-    private static final Color VALUE_NORMAL_COLOR = Color.GREEN;
-    private static final Color VALUE_TOO_LOW_COLOR = Color.CYAN;
-    protected static final double[] temperatureBounds = new double[]{-100,
-            60};  // too low, too high values
-    protected static final double[] voltageBounds = new double[]{4000.0,
+public class Cell extends NumAlerter {
+    public static final double[] VOLTAGE_BOUNDS = new double[]{4000.0,
             4200.0};
-    protected final JLabel temperatureLabel = new JLabel("0000.00");
     // labels
     protected final JLabel voltageLabel = new JLabel("0000.00");
-    protected double temperature;  // remember last value set
     protected double voltage;
+    private int indexInBms;
 
-    Cell() {
+    Cell(int indexInBms) {
+        this.indexInBms = indexInBms;
         setLayout(new GridLayout(2, 1));  // rows, columns
         setup();
     }
@@ -49,25 +41,6 @@ class Cell extends JPanel {
     /*
      * Getters & setters
      */
-
-    /**
-     * Retrieve temperature value
-     *
-     * @return getter method for temperature
-     */
-    public double getTemperature() {
-        return temperature;
-    }
-
-    /**
-     * Show new value in label and bar
-     *
-     * @param value new temperature to set
-     */
-    public void setTemperature(double value) {
-        temperature = value;  // update value
-        update(temperatureLabel, temperatureBounds, value);
-    }
 
     /**
      * Retrieve voltage value
@@ -85,22 +58,7 @@ class Cell extends JPanel {
      */
     public void setVoltage(double value) {
         voltage = value;  // update value
-        update(voltageLabel, voltageBounds, value);
-    }
-
-    /**
-     * Sets temperature bounds
-     *
-     * @param min minimum value; below this value background will color
-     *            VALUE_TOO_LOW_COLOR
-     * @param max maximum value; over this value background will color
-     *            VALUE_TOO_HIGH_COLOR
-     */
-    void setTemperatureBounds(double min, double max) {
-        temperatureBounds[0] = min;
-        temperatureBounds[1] = max;
-
-        setTemperature(temperature);  // update colors
+        update(voltageLabel, VOLTAGE_BOUNDS, value);
     }
 
     /**
@@ -112,19 +70,10 @@ class Cell extends JPanel {
      *            VALUE_TOO_HIGH_COLOR
      */
     void setVoltageBounds(double min, double max) {
-        voltageBounds[0] = min;
-        voltageBounds[1] = max;
+        VOLTAGE_BOUNDS[0] = min;
+        VOLTAGE_BOUNDS[1] = max;
 
         setVoltage(voltage);  // update colors
-    }
-
-    /**
-     * Get temperature bounds
-     *
-     * @return [min, max] array
-     */
-    double[] getTemperatureBounds() {
-        return temperatureBounds;
     }
 
     /**
@@ -133,7 +82,7 @@ class Cell extends JPanel {
      * @return [min, max] array
      */
     double[] getVoltageBounds() {
-        return voltageBounds;
+        return VOLTAGE_BOUNDS;
     }
 
     /**
@@ -148,7 +97,6 @@ class Cell extends JPanel {
 
         Timer updater = new Timer(10, e -> {
             dialog.updateOrFail(0, getVoltage());
-            dialog.updateOrFail(1, getTemperature());
         });  // timer to update dialog values
         updater.start();
     }
@@ -161,40 +109,11 @@ class Cell extends JPanel {
      * Add to panel all widgets
      */
     private void setup() {
-        temperatureLabel.setHorizontalAlignment(SwingConstants.RIGHT);
         voltageLabel.setHorizontalAlignment(SwingConstants.RIGHT);
 
-        add(temperatureLabel);  // add items
+        add(new JLabel(Integer.toString(indexInBms)));
         add(voltageLabel);
 
         setVisible(true);  // open
-    }
-
-    /**
-     * Update widget with new value
-     *
-     * @param label  label to update series
-     * @param bounds min, max of value allowed
-     * @param value  new value
-     */
-    protected void update(JLabel label, double[] bounds, double value) {
-        label.setText(SHORT_DEC_FORMAT.format(value));
-        updateBackground(value, bounds);
-    }
-
-    /**
-     * Changes background on given update series of value
-     *
-     * @param value  new value to update
-     * @param bounds min, max of value
-     */
-    private void updateBackground(double value, double[] bounds) {
-        if (value >= bounds[1]) {  // too high
-            setBackground(VALUE_TOO_HIGH_COLOR);
-        } else if (value <= bounds[0]) {  // too low
-            setBackground(VALUE_TOO_LOW_COLOR);
-        } else {  // normal
-            setBackground(VALUE_NORMAL_COLOR);
-        }
     }
 }